se creo un main fragment para tener un contendeor de los demas fragments, ya que sino es mas dificl el control con el splash y un activity teniendo el control sobre todos los fragmentos.(Falta mejroar mas el splash)
