+++ app/src/main/res/drawable/ic_icon_app.xml
@@ -1,0 +1,45 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="800dp"
+    android:height="800dp"
+    android:viewportWidth="512"
+    android:viewportHeight="512">
+  <path
+      android:pathData="M256,256m-256,0a256,256 0,1 1,512 0a256,256 0,1 1,-512 0"
+      android:fillColor="#324A5E"/>
+  <path
+      android:pathData="M508.83,296.32l-97.2,-97.2L165.42,301.61l-51.64,17.98l188.27,188.27C408.42,488.54 491.87,403.49 508.83,296.32z"
+      android:fillColor="#2B3B4E"/>
+  <path
+      android:pathData="M348.03,170.67H170.67c-47.13,0 -85.33,38.21 -85.33,85.33s38.21,85.33 85.33,85.33c25.27,0 47.96,-10.99 63.59,-28.44h50.19c15.63,17.45 38.32,28.44 63.59,28.44c47.13,0 85.33,-38.21 85.33,-85.33S395.16,170.67 348.03,170.67z"
+      android:fillColor="#E6F3FF"/>
+  <path
+      android:pathData="M348.03,170.67h-91.46v142.22h27.87c15.63,17.45 38.32,28.44 63.59,28.44c47.13,0 85.33,-38.21 85.33,-85.33S395.16,170.67 348.03,170.67z"
+      android:fillColor="#CFDBE6"/>
+  <path
+      android:pathData="M170.67,256m-56.89,0a56.89,56.89 0,1 1,113.78 0a56.89,56.89 0,1 1,-113.78 0"
+      android:fillColor="#CFDBE6"/>
+  <path
+      android:pathData="M348.04,256m-56.89,0a56.89,56.89 0,1 1,113.78 0a56.89,56.89 0,1 1,-113.78 0"
+      android:fillColor="#B8C9D9"/>
+  <path
+      android:pathData="M194.1,246.52h-8.94c-2.77,0 -5.01,-2.24 -5.01,-5.01v-8.94c0,-2.77 -2.24,-5.01 -5.01,-5.01h-8.94c-2.77,0 -5.01,2.24 -5.01,5.01v8.94c0,2.77 -2.24,5.01 -5.01,5.01h-8.94c-2.77,0 -5.01,2.24 -5.01,5.01v8.94c0,2.77 2.24,5.01 5.01,5.01h8.94c2.77,0 5.01,2.24 5.01,5.01v8.94c0,2.77 2.24,5.01 5.01,5.01h8.94c2.77,0 5.01,-2.24 5.01,-5.01v-8.94c0,-2.77 2.24,-5.01 5.01,-5.01h8.94c2.77,0 5.01,-2.24 5.01,-5.01v-8.94C199.11,248.76 196.87,246.52 194.1,246.52z"
+      android:fillColor="#778DA0"/>
+  <path
+      android:pathData="M348.04,234.61m-12.22,0a12.22,12.22 0,1 1,24.45 0a12.22,12.22 0,1 1,-24.45 0"
+      android:fillColor="#4CDBC4"/>
+  <path
+      android:pathData="M348.04,277.39m-12.22,0a12.22,12.22 0,1 1,24.45 0a12.22,12.22 0,1 1,-24.45 0"
+      android:fillColor="#FFFFFF"/>
+  <path
+      android:pathData="M369.42,256m-12.22,0a12.22,12.22 0,1 1,24.45 0a12.22,12.22 0,1 1,-24.45 0"
+      android:fillColor="#FD8469"/>
+  <path
+      android:pathData="M326.65,256m-12.22,0a12.22,12.22 0,1 1,24.45 0a12.22,12.22 0,1 1,-24.45 0"
+      android:fillColor="#FF80AA"/>
+  <path
+      android:pathData="M218.94,206.16L218.94,206.16c0,-4.5 3.65,-8.15 8.15,-8.15h57.83c4.5,0 8.15,3.65 8.15,8.15l0,0c0,4.5 -3.65,8.15 -8.15,8.15h-57.83C222.58,214.31 218.94,210.66 218.94,206.16z"
+      android:fillColor="#CFDBE6"/>
+  <path
+      android:pathData="M284.92,198.01h-28.34v16.3h28.34c4.5,0 8.15,-3.65 8.15,-8.15C293.06,201.66 289.42,198.01 284.92,198.01z"
+      android:fillColor="#B8C9D9"/>
+</vector>
