+++ app/src/main/res/navigation/fragment_main_nav_graph.xml
@@ -17,11 +17,6 @@
         tools:layout="@layout/fragment_calculator" />
 
     <fragment
-        android:id="@+id/splashFragment"
-        android:name="com.app.gamercalculator.ui.views.fragments.SplashFragment"
-        android:label="fragment_splash"
-        tools:layout="@layout/fragment_splash" />
-    <fragment
         android:id="@+id/subscriptionsFragment"
         android:name="com.app.gamercalculator.ui.views.fragments.SubscriptionsFragment"
         android:label="fragment_subscriptions"
