+++ app/src/main/res/values/colors.xml
@@ -8,4 +8,5 @@
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
     <color name="gray">#80747474</color>
+    <color name="gray_obscure">#E51C1C38</color>
 </resources>
